Feat: 아이디, 닉네임 중복체크 API 생성, JWT, Refresh Token 발급 구현 Docs: Swagger 사용법 주석 작성

+++ src/main/java/com/example/demo/dto/member/MemberDetails.java
@@ -17,7 +17,7 @@ public class MemberDetails implements UserDetails {
     public String getNickname() {
         return member.getNickname();
     }
-    public short getAge() { return member.getAge(); }
+    public Integer getAge() { return member.getAge(); }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
